Moved SymPyDeprecationWarning
