+++ sympy/core/decorators.py
@@ -5,9 +5,9 @@ The purpose of this module is to expose decorators without any other
 dependencies, so that they can be easily imported anywhere in sympy/core.
 """
 
-from sympify import SympifyError, sympify
 from functools import wraps
-from sympify import SymPyDeprecationWarning, SympifyError, sympify
+from sympify import SympifyError, sympify
+from sympy.core.compatibility import SymPyDeprecationWarning
 import warnings
 
 def deprecated(func):
